Fix (MSVC) compiler warnings reg. `size_t` vs. `int`
as per code review, https://github.com/musescore/MuseScore/pull/11197#pullrequestreview-943197231

+++ src/engraving/libmscore/spanner.h
@@ -62,7 +62,7 @@ public:
     // Score Tree functions
     virtual EngravingObject* scanParent() const override;
     virtual EngravingObject* scanChild(int idx) const override;
-    virtual int scanChildCount() const override;
+    virtual size_t scanChildCount() const override;
 
     virtual qreal mag() const override;
     virtual Fraction tick() const override;
@@ -188,7 +188,7 @@ public:
     // Score Tree functions
     virtual EngravingObject* scanParent() const override;
     virtual EngravingObject* scanChild(int idx) const override;
-    virtual int scanChildCount() const override;
+    virtual size_t scanChildCount() const override;
 
     virtual qreal mag() const override;
